refactor(runtime): split builtins_system.rs into cohesive modules (§7-8)
Split the 1938-line builtins_system.rs (over the 500-line limit) into
five cohesive modules, all under 500 lines:

- builtins_system.rs (193)         - module-level proc/pipe state maps,
                                      ProcOptions, spawn_callable_promise
- builtins_system_require.rs (457) - require/module-load + symbol import
- builtins_system_proc.rs (339)    - proc/pipe construction, run-arg
                                      option parsing, qx (+ unit tests)
- builtins_system_run.rs (469)     - builtin_run / shell
- builtins_system_async.rs (494)   - sleep/start/await/kill/signal

All methods are impl Interpreter, so the split is visibility-only. The
module-level state-map fns and ProcOptions stay in the main module (kept
to preserve the crate::runtime::builtins_system::* paths used by
methods_promise / native_io / supply_transform); the proc helpers shared
with builtin_run are bumped to pub(super), and sibling modules glob in
super::builtins_system::* where needed. No behavior change.
