+++ src/runtime/builtins_system.rs
@@ -86,20 +86,20 @@ pub(crate) fn proc_by_pid_map() -> &'static ProcByPidMap {
 }
 
 /// Options extracted from named arguments for run/shell.
-struct ProcOptions {
-    cwd: Option<String>,
-    env: HashMap<String, String>,
-    capture_err: bool,
-    capture_out: bool,
-    capture_in: bool,
-    in_pipe_pid: Option<i64>,
-    in_pipe_content: Option<String>,
-    bin: bool,
-    win_verbatim_args: bool,
+pub(super) struct ProcOptions {
+    pub(super) cwd: Option<String>,
+    pub(super) env: HashMap<String, String>,
+    pub(super) capture_err: bool,
+    pub(super) capture_out: bool,
+    pub(super) capture_in: bool,
+    pub(super) in_pipe_pid: Option<i64>,
+    pub(super) in_pipe_content: Option<String>,
+    pub(super) bin: bool,
+    pub(super) win_verbatim_args: bool,
     /// When :out is given an IO::Handle, store the handle ID to dup the fd
-    out_handle_id: Option<usize>,
+    pub(super) out_handle_id: Option<usize>,
     /// When :merge is True, redirect stderr to stdout
-    merge: bool,
+    pub(super) merge: bool,
 }
 
 impl Interpreter {
@@ -189,1750 +189,5 @@ impl Interpreter {
         });
 
         ret
-    }
-
-    fn missing_symbol_name_from_failure(value: &Value) -> Option<String> {
-        let Value::Instance {
-            class_name,
-            attributes,
-            ..
-        } = value
-        else {
-            return None;
-        };
-        if class_name != "Failure" {
-            return None;
-        }
-        let map = attributes.as_map();
-        let exception = map.get("exception")?;
-        let Value::Instance {
-            attributes: ex_attrs,
-            ..
-        } = exception
-        else {
-            return None;
-        };
-        let message = ex_attrs.as_map().get("message")?.to_string_value();
-        let prefix = "No such symbol '";
-        let rest = message.strip_prefix(prefix)?;
-        let symbol = rest.strip_suffix('\'')?;
-        if symbol.is_empty() {
-            None
-        } else {
-            Some(symbol.to_string())
-        }
-    }
-
-    fn require_target_is_path_like(target: &str) -> bool {
-        target.ends_with(".rakumod")
-            || target.ends_with(".pm6")
-            || target.contains('/')
-            || target.contains('\\')
-    }
-
-    fn require_guess_module_name_from_path(path: &str) -> Option<String> {
-        let p = Path::new(path);
-        let stem = p.file_stem()?.to_string_lossy().to_string();
-        if stem.is_empty() { None } else { Some(stem) }
-    }
-
-    fn resolve_require_file_path(&self, file: &str) -> Option<PathBuf> {
-        let direct = PathBuf::from(file);
-        if direct.exists() {
-            return Some(direct);
-        }
-        for base in &self.lib_paths {
-            let candidate = Path::new(base).join(file);
-            if candidate.exists() {
-                return Some(candidate);
-            }
-            let candidate = Path::new(base).join("lib").join(file);
-            if candidate.exists() {
-                return Some(candidate);
-            }
-        }
-        if let Some(cwd) = self.get_dynamic_string("$*CWD") {
-            let candidate = Path::new(&cwd).join(file);
-            if candidate.exists() {
-                return Some(candidate);
-            }
-        }
-        None
-    }
-
-    /// Parse a source file for require, returning the merged AST.
-    fn parse_require_source(
-        &mut self,
-        file: &str,
-        path: &std::path::Path,
-    ) -> Result<Vec<crate::ast::Stmt>, RuntimeError> {
-        let code = std::fs::read_to_string(path)
-            .map_err(|e| RuntimeError::new(format!("Failed to read module {}: {}", file, e)))?;
-        let preprocessed = Self::maybe_preprocess_roast_directives(&code);
-        crate::parser::set_parser_lib_paths(self.lib_paths.clone());
-        crate::parser::set_parser_program_path(self.program_path.clone());
-        let result = parse_dispatch::parse_source(&preprocessed);
-        crate::parser::clear_parser_lib_paths();
-        for warning in crate::parser::take_parse_warnings() {
-            self.write_warn_to_stderr(&warning);
-        }
-        let stmts = result.map(|(stmts, _)| stmts).map_err(|mut err| {
-            err.message = format!("Failed to parse module '{}': {}", file, err.message);
-            err
-        })?;
-        Ok(Self::merge_unit_class(stmts))
-    }
-
-    fn require_load_from_file(
-        &mut self,
-        file: &str,
-        package_hint: Option<&str>,
-    ) -> Result<(), RuntimeError> {
-        let path = self
-            .resolve_require_file_path(file)
-            .ok_or_else(|| RuntimeError::unsatisfied_dependency(file))?;
-
-        // Try loading from precompilation cache
-        let stmts = if self.precomp_enabled {
-            if let Some(cached) = crate::precomp::load_cached_ast(&path) {
-                cached
-            } else {
-                let parsed = self.parse_require_source(file, &path)?;
-                crate::precomp::save_cached_ast(&path, &parsed);
-                parsed
-            }
-        } else {
-            self.parse_require_source(file, &path)?
-        };
-        let saved_package = self.current_package();
-        let before_function_keys: std::collections::HashSet<Symbol> =
-            self.registry().functions.keys().copied().collect();
-        let before_env_keys: std::collections::HashSet<Symbol> = self.env.keys().copied().collect();
-        let before_class_keys: std::collections::HashSet<String> =
-            self.registry().classes.keys().cloned().collect();
-        if let Some(pkg) = package_hint
-            && !pkg.is_empty()
-        {
-            self.set_current_package(pkg.to_string());
-        }
-        let run_result = self.run_block(&stmts);
-        self.set_current_package(saved_package);
-        run_result?;
-
-        // A `sub MAIN` defined in a required module is NOT the program's MAIN
-        // and must not be auto-dispatched at program end. Remove any top-level
-        // MAIN routines that this module introduced.
-        Self::remove_leaked_main_routines(
-            &mut self.registry_mut().functions,
-            &before_function_keys,
-        );
-
-        if let Some(pkg) = package_hint
-            && !pkg.is_empty()
-        {
-            let mut fn_aliases: Vec<(Symbol, std::sync::Arc<FunctionDef>)> = Vec::new();
-            for (name, def) in &self.registry().functions {
-                if before_function_keys.contains(name) {
-                    continue;
-                }
-                let name_s = name.resolve();
-                let tail = name_s.rsplit_once("::").map(|(_, t)| t).unwrap_or(&name_s);
-                if tail.contains('/') || tail.contains(':') {
-                    continue;
-                }
-                let alias = format!("{pkg}::{tail}");
-                let alias_sym = Symbol::intern(&alias);
-                if !self.registry().functions.contains_key(&alias_sym) {
-                    fn_aliases.push((alias_sym, def.clone()));
-                }
-            }
-            for (alias, def) in fn_aliases {
-                self.registry_mut().functions.insert(alias, def);
-            }
-
-            let mut env_aliases: Vec<(String, Value)> = Vec::new();
-            for (name, value) in &self.env {
-                if before_env_keys.contains(name) {
-                    continue;
-                }
-                if name.starts_with("$") || name.starts_with("@") || name.starts_with("%") {
-                    continue;
-                }
-                let name_s = name.resolve();
-                let tail = name_s.rsplit_once("::").map(|(_, t)| t).unwrap_or(&name_s);
-                let alias = format!("{pkg}::{tail}");
-                if !self.env.contains_key(&alias) {
-                    env_aliases.push((alias, value.clone()));
-                }
-            }
-            for (alias, value) in env_aliases {
-                self.env.insert(alias, value);
-            }
-
-            let mut class_aliases: Vec<(String, ClassDef)> = Vec::new();
-            for (name, class_def) in &self.registry().classes {
-                if before_class_keys.contains(name) {
-                    continue;
-                }
-                let tail = name.rsplit_once("::").map(|(_, t)| t).unwrap_or(name);
-                let alias = format!("{pkg}::{tail}");
-                if !self.registry().classes.contains_key(&alias) {
-                    class_aliases.push((alias, class_def.clone()));
-                }
-            }
-            for (alias, class_def) in class_aliases {
-                self.registry_mut().classes.insert(alias, class_def);
-            }
-        }
-
-        Ok(())
-    }
-
-    /// Remove top-level `MAIN` routines that were registered while loading a
-    /// module (via `require`/`use`). A `sub MAIN` declared in a loaded module
-    /// is not the program's MAIN and must not be auto-dispatched at program
-    /// end. `before_keys` is the set of function-registry keys that existed
-    /// before the module body ran; only newly-added MAIN keys are removed.
-    pub(crate) fn remove_leaked_main_routines(
-        functions: &mut HashMap<Symbol, std::sync::Arc<FunctionDef>>,
-        before_keys: &std::collections::HashSet<Symbol>,
-    ) {
-        let leaked: Vec<Symbol> = functions
-            .keys()
-            .filter(|k| {
-                if before_keys.contains(*k) {
-                    return false;
-                }
-                let ks = k.resolve();
-                // Match GLOBAL::MAIN, GLOBAL::MAIN/<arity>, GLOBAL::MAIN/<...>:<sig>
-                // i.e. the routine's short name is exactly `MAIN`.
-                let after_pkg = ks.rsplit_once("::").map(|(_, t)| t).unwrap_or(&ks);
-                let short = after_pkg.split(['/', ':']).next().unwrap_or(after_pkg);
-                short == "MAIN"
-            })
-            .copied()
-            .collect();
-        for k in leaked {
-            functions.remove(&k);
-        }
-    }
-
-    fn import_single_require_symbol(&mut self, module: &str, symbol: &str) -> bool {
-        if symbol.is_empty() {
-            return true;
-        }
-        if module == "Test" {
-            return true;
-        }
-        if let Some(name) = symbol.strip_prefix('&') {
-            let source_single = format!("{module}::{name}");
-            let source_prefix = format!("{module}::{name}/");
-            let target_single = format!("GLOBAL::{name}");
-            let target_prefix = format!("GLOBAL::{name}/");
-            let mut found = false;
-            let function_entries: Vec<(Symbol, std::sync::Arc<FunctionDef>)> = self
-                .registry()
-                .functions
-                .iter()
-                .filter_map(|(k, v)| {
-                    let ks = k.resolve();
-                    if ks == source_single {
-                        found = true;
-                        Some((Symbol::intern(&target_single), v.clone()))
-                    } else if ks.starts_with(&source_prefix) {
-                        found = true;
-                        Some((
-                            Symbol::intern(&ks.replacen(&source_prefix, &target_prefix, 1)),
-                            v.clone(),
-                        ))
-                    } else {
-                        None
-                    }
-                })
-                .collect();
-            for (k, v) in function_entries {
-                self.registry_mut().functions.insert(k, v);
-            }
-            return found
-                || self
-                    .registry()
-                    .functions
-                    .contains_key(&Symbol::intern(&format!("GLOBAL::{name}")));
-        }
-
-        if let Some(sigil) = symbol.chars().next()
-            && matches!(sigil, '$' | '@' | '%' | '&')
-        {
-            let bare = &symbol[1..];
-            let candidates = [
-                format!("{sigil}{module}::{bare}"),
-                format!("{sigil}{bare}"),
-                format!("{module}::{bare}"),
-                bare.to_string(),
-            ];
-            for source in candidates {
-                if let Some(value) = self.env.get(&source).cloned() {
-                    self.env.insert(symbol.to_string(), value);
-                    return true;
-                }
-            }
-            return false;
-        }
-
-        let source_single = format!("{module}::{symbol}");
-        if self.has_class(&source_single) || self.is_role(&source_single) {
-            self.env.insert(
-                symbol.to_string(),
-                Value::Package(Symbol::intern(&source_single)),
-            );
-            self.env.insert(
-                format!("__mutsu_sigilless_readonly::{symbol}"),
-                Value::Bool(true),
-            );
-            return true;
-        }
-        if self.has_class(symbol) || self.is_role(symbol) {
-            self.env
-                .insert(symbol.to_string(), Value::Package(Symbol::intern(symbol)));
-            self.env.insert(
-                format!("__mutsu_sigilless_readonly::{symbol}"),
-                Value::Bool(true),
-            );
-            return true;
-        }
-        if let Some(value) = self.env.get(&source_single).cloned() {
-            if self
-                .registry()
-                .functions
-                .contains_key(&Symbol::intern(&source_single))
-                && matches!(value, Value::Int(_))
-            {
-                return false;
-            }
-            self.env.insert(symbol.to_string(), value);
-            self.env.insert(
-                format!("__mutsu_sigilless_readonly::{symbol}"),
-                Value::Bool(true),
-            );
-            return true;
-        }
-        if let Some(value) = self.env.get(symbol).cloned() {
-            let is_nil = matches!(value, Value::Nil);
-            self.env.insert(symbol.to_string(), value);
-            self.env.insert(
-                format!("__mutsu_sigilless_readonly::{symbol}"),
-                Value::Bool(true),
-            );
-            return !is_nil;
-        }
-        false
-    }
-
-    fn require_import_symbols(
-        &mut self,
-        module: &str,
-        imports: &[String],
-    ) -> Result<(), RuntimeError> {
-        let mut missing = Vec::new();
-        for symbol in imports {
-            if !self.import_single_require_symbol(module, symbol) {
-                missing.push(symbol.clone());
-            }
-        }
-        if missing.is_empty() {
-            Ok(())
-        } else {
-            Err(RuntimeError::new(format!(
-                "X::Import::MissingSymbols: {}",
-                missing.join(" ")
-            )))
-        }
-    }
-
-    pub(super) fn builtin_require(&mut self, args: &[Value]) -> Result<Value, RuntimeError> {
-        let mut module_value: Option<Value> = None;
-        let mut file_target: Option<String> = None;
-        let mut imports: Vec<String> = Vec::new();
-        for arg in args {
-            match arg {
-                Value::Pair(key, value) if key == "__mutsu_require_file" => {
-                    file_target = Some(value.to_string_value());
-                }
-                Value::Array(items, ..) => {
-                    imports.extend(items.iter().map(|v| v.to_string_value()));
-                }
-                other => {
-                    if module_value.is_none() {
-                        module_value = Some(other.clone());
-                    } else {
-                        imports.push(other.to_string_value());
-                    }
-                }
-            }
-        }
-        let module_value =
-            module_value.ok_or_else(|| RuntimeError::new("require expects a module"))?;
-        let module_string = match &module_value {
-            Value::Package(name) => name.resolve(),
-            Value::Str(name) => name.to_string(),
-            value @ Value::Instance { .. } => Self::missing_symbol_name_from_failure(value)
-                .unwrap_or_else(|| value.to_string_value()),
-            other => other.to_string_value(),
-        };
-        let return_is_str = matches!(module_value, Value::Str(_));
-        let mut module_name = if module_string.is_empty() {
-            None
-        } else {
-            Some(module_string.clone())
-        };
-
-        let load_from_file = file_target.is_some()
-            || module_name
-                .as_ref()
-                .is_some_and(|m| Self::require_target_is_path_like(m));
-        if load_from_file {
-            let file = file_target
-                .clone()
-                .or_else(|| module_name.clone())
-                .ok_or_else(|| RuntimeError::new("require expects a file"))?;
-            if file_target.is_none() {
-                module_name = Self::require_guess_module_name_from_path(&file);
-            }
-            self.require_load_from_file(&file, module_name.as_deref())?;
-            if let Some(module) = module_name.as_ref() {
-                self.loaded_modules.insert(module.clone());
-            }
-        } else if let Some(module) = module_name.as_ref() {
-            // For `require`, force a full reload if the module was previously loaded
-            // but its functions were cleaned up by pop_import_scope (e.g. when
-            // a previous `require` happened inside a block that has since exited).
-            let prefix = format!("{module}::");
-            if self.loaded_modules.contains(module)
-                && !self
-                    .registry()
-                    .functions
-                    .keys()
-                    .any(|k| k.resolve().starts_with(&prefix))
-            {
-                self.loaded_modules.remove(module);
-            }
-            self.use_module(module)?;
-        } else {
-            return Err(RuntimeError::new("require expects a module name"));
-        }
-
-        let in_method_context = !self.method_class_stack.is_empty();
-        let should_install_stub = !return_is_str || !in_method_context;
-        if let Some(module) = module_name.as_ref()
-            && should_install_stub
-        {
-            self.env
-                .insert(module.clone(), Value::Package(Symbol::intern(module)));
-            if !imports.is_empty() {
-                self.require_import_symbols(module, &imports)?;
-            }
-        } else if let Some(module) = module_name.as_ref()
-            && !imports.is_empty()
-        {
-            self.require_import_symbols(module, &imports)?;
-        }
-
-        if return_is_str {
-            Ok(Value::str(module_string))
-        } else {
-            let name = module_name.unwrap_or(module_string);
-            Ok(Value::Package(Symbol::intern(&name)))
-        }
-    }
-
-    pub(super) fn builtin_gethost(&self, args: &[Value]) -> Result<Value, RuntimeError> {
-        let host_str = args.first().map(|v| v.to_string_value());
-        let hostname = host_str.unwrap_or_else(Self::hostname);
-        let addrs = Self::resolve_host(&hostname);
-        Ok(Self::make_os_name_value(hostname, addrs))
-    }
-
-    pub(super) fn builtin_chroot(&mut self, args: &[Value]) -> Result<Value, RuntimeError> {
-        let path_str = args
-            .first()
-            .map(|v| v.to_string_value())
-            .or_else(|| {
-                self.get_dynamic_string("$*CWD")
-                    .or_else(|| self.get_dynamic_string("$*CHROOT"))
-            })
-            .unwrap_or_else(|| ".".to_string());
-        let path_buf = PathBuf::from(path_str.clone());
-        if !path_buf.is_dir() {
-            return Ok(Value::Bool(false));
-        }
-        let canonical = path_buf.canonicalize().unwrap_or_else(|_| path_buf.clone());
-        if std::env::set_current_dir(&canonical).is_err() {
-            return Ok(Value::Bool(false));
-        }
-        self.chroot_root = Some(canonical.clone());
-        let repr = Self::stringify_path(&canonical);
-        self.env
-            .insert("$*CHROOT".to_string(), Value::str(repr.clone()));
-        self.env
-            .insert("*CHROOT".to_string(), Value::str(repr.clone()));
-        let cwd_val = self.make_io_path_instance(&repr);
-        self.env.insert("$*CWD".to_string(), cwd_val.clone());
-        self.env.insert("*CWD".to_string(), cwd_val);
-        Ok(Value::Bool(true))
-    }
-
-    /// Helper to create a Proc instance from process execution results.
-    fn make_proc_instance(
-        exitcode: i64,
-        signal: i64,
-        pid: i64,
-        command: Value,
-        captured_err: Option<String>,
-        captured_out: Option<String>,
-    ) -> Value {
-        Self::make_proc_instance_bin(
-            exitcode,
-            signal,
-            pid,
-            command,
-            captured_err,
-            captured_out,
-            false,
-        )
-    }
-
-    fn make_proc_instance_bin(
-        exitcode: i64,
-        signal: i64,
-        pid: i64,
-        command: Value,
-        captured_err: Option<String>,
-        captured_out: Option<String>,
-        bin: bool,
-    ) -> Value {
-        let mut attrs = HashMap::new();
-        attrs.insert("exitcode".to_string(), Value::Int(exitcode));
-        attrs.insert("signal".to_string(), Value::Int(signal));
-        attrs.insert("pid".to_string(), Value::Int(pid));
-        attrs.insert("command".to_string(), command);
-        if bin {
-            attrs.insert("bin".to_string(), Value::Bool(true));
-        }
-        let mut pipe_ids = Vec::new();
-        if let Some(err_content) = captured_err {
-            let pipe = Self::make_io_pipe_bin(err_content, bin);
-            if let Value::Instance {
-                attributes: ref a, ..
-            } = pipe
-                && let Some(Value::Int(id)) = a.as_map().get("pipe-id")
-            {
-                pipe_ids.push(*id);
-            }
-            attrs.insert("err".to_string(), pipe);
-        }
-        if let Some(out_content) = captured_out {
-            let pipe = Self::make_io_pipe_bin(out_content, bin);
-            if let Value::Instance {
-                attributes: ref a, ..
-            } = pipe
-                && let Some(Value::Int(id)) = a.as_map().get("pipe-id")
-            {
-                pipe_ids.push(*id);
-            }
-            attrs.insert("out".to_string(), pipe);
-        }
-        let proc_val = Value::make_instance(Symbol::intern("Proc"), attrs);
-        if !pipe_ids.is_empty()
-            && let Ok(mut map) = pipe_proc_map().lock()
-        {
-            for id in pipe_ids {
-                map.insert(id, proc_val.clone());
-            }
-        }
-        proc_val
-    }
-
-    /// Create an IO::Pipe instance wrapping captured content.
-    pub(super) fn make_io_pipe(content: String) -> Value {
-        Self::make_io_pipe_bin(content, false)
-    }
-
-    fn make_io_pipe_bin(content: String, bin: bool) -> Value {
-        use std::sync::atomic::{AtomicI64, Ordering};
-        static NEXT_PIPE_ID: AtomicI64 = AtomicI64::new(1);
-        let id = NEXT_PIPE_ID.fetch_add(1, Ordering::SeqCst);
-        if let Ok(mut map) = io_pipe_state_map().lock() {
-            map.insert(
-                id,
-                IoPipeState {
-                    content: content.clone(),
-                    cursor: 0,
-                    closed: false,
-                },
-            );
-        }
-        let mut attrs = HashMap::new();
-        attrs.insert("content".to_string(), Value::str(content));
-        attrs.insert("pipe-id".to_string(), Value::Int(id));
-        if bin {
-            attrs.insert("bin".to_string(), Value::Bool(true));
-        }
-        Value::make_instance(Symbol::intern("IO::Pipe"), attrs)
-    }
-
-    /// Extract named options (cwd, env, :err, :out) from arguments.
-    fn extract_proc_options(args: &[Value], skip: usize) -> ProcOptions {
-        let mut opts = ProcOptions {
-            cwd: None,
-            env: HashMap::new(),
-            capture_err: false,
-            capture_out: false,
-            capture_in: false,
-            in_pipe_pid: None,
-            in_pipe_content: None,
-            bin: false,
-            win_verbatim_args: false,
-            out_handle_id: None,
-            merge: false,
-        };
-        for value in args.iter().skip(skip) {
-            if let Value::Hash(map) = value {
-                for (key, inner) in map.iter() {
-                    match key.as_str() {
-                        "cwd" => {
-                            opts.cwd = Some(inner.to_string_value());
-                        }
-                        "env" => {
-                            if let Value::Hash(env_map) = inner {
-                                for (ek, ev) in env_map.iter() {
-                                    opts.env.insert(ek.clone(), ev.to_string_value());
-                                }
-                            }
-                        }
-                        "err" => {
-                            opts.capture_err = inner.truthy();
-                        }
-                        "out" => {
-                            Self::extract_out_option(inner, &mut opts);
-                        }
-                        "in" => {
-                            Self::extract_in_option(inner, &mut opts);
-                        }
-                        "bin" => {
-                            opts.bin = inner.truthy();
-                        }
-                        "merge" => {
-                            opts.merge = inner.truthy();
-                        }
-                        "win-verbatim-args" => {
-                            opts.win_verbatim_args = inner.truthy();
-                        }
-                        _ => {}
-                    }
-                }
-            }
-            if let Value::Pair(key, inner) = value {
-                match key.as_str() {
-                    "cwd" => opts.cwd = Some(inner.to_string_value()),
-                    "env" => {
-                        if let Value::Hash(env_map) = inner.as_ref() {
-                            for (ek, ev) in env_map.iter() {
-                                opts.env.insert(ek.clone(), ev.to_string_value());
-                            }
-                        }
-                    }
-                    "err" => opts.capture_err = inner.truthy(),
-                    "out" => Self::extract_out_option(inner, &mut opts),
-                    "in" => Self::extract_in_option(inner, &mut opts),
-                    "bin" => opts.bin = inner.truthy(),
-                    "merge" => opts.merge = inner.truthy(),
-                    "win-verbatim-args" => opts.win_verbatim_args = inner.truthy(),
-                    _ => {}
-                }
-            }
-        }
-        opts
-    }
-
-    fn extract_in_option(value: &Value, opts: &mut ProcOptions) {
-        if let Value::Instance {
-            class_name,
-            attributes,
-            ..
-        } = value
-            && class_name.resolve() == "IO::Pipe"
-        {
-            opts.capture_in = true;
-            if let Some(Value::Int(pid)) = attributes.as_map().get("live-pid") {
-                opts.in_pipe_pid = Some(*pid);
-                return;
-            }
-            if let Some(Value::Int(pid)) = attributes.as_map().get("proc-pid") {
-                opts.in_pipe_pid = Some(*pid);
-                return;
-            }
-            let content = if let Some(Value::Int(id)) = attributes.as_map().get("pipe-id")
-                && let Ok(mut map) = io_pipe_state_map().lock()
-                && let Some(state) = map.get_mut(id)
-            {
-                let c = state.content[state.cursor..].to_string();
-                state.cursor = state.content.len();
-                c
-            } else {
-                attributes
-                    .as_map()
-                    .get("content")
-                    .map(|v| v.to_string_value())
-                    .unwrap_or_default()
-            };
-            opts.in_pipe_content = Some(content);
-            return;
-        }
-        opts.capture_in = value.truthy();
-    }
-
-    fn extract_out_option(value: &Value, opts: &mut ProcOptions) {
-        // :out can be an IO::Handle (redirect stdout to that file handle)
-        if let Value::Instance {
-            class_name,
-            attributes,
-            ..
-        } = value
-            && class_name.resolve() == "IO::Handle"
-            && let Some(Value::Int(id)) = attributes.as_map().get("handle")
-        {
-            opts.out_handle_id = Some(*id as usize);
-            return;
-        }
-        opts.capture_out = value.truthy();
-    }
-
-    #[cfg(windows)]
-    fn apply_run_args(cmd: &mut Command, args: &[String], win_verbatim_args: bool) {
-        use std::os::windows::process::CommandExt;
-
-        if win_verbatim_args {
-            if !args.is_empty() {
-                // Raku's :win-verbatim-args forwards argv payload without Win32 quoting.
-                cmd.raw_arg(args.join(" "));
-            }
-        } else {
-            cmd.args(args);
-        }
-    }
-
-    #[cfg(not(windows))]
-    fn apply_run_args(cmd: &mut Command, args: &[String], _win_verbatim_args: bool) {
-        cmd.args(args);
-    }
-
-    pub(super) fn builtin_run(&self, args: &[Value]) -> Result<Value, RuntimeError> {
-        if args.is_empty() {
-            return Ok(Self::make_proc_instance(
-                -1,
-                0,
-                0,
-                Value::str(String::new()),
-                None,
-                None,
-            ));
-        }
-
-        // Collect positional string args and named options
-        let mut positional: Vec<String> = Vec::new();
-        let mut first_arg_io_path = false;
-        for (idx, arg) in args.iter().enumerate() {
-            let arg = match arg {
-                Value::Capture { named, .. }
-                    if matches!(named.get("__mutsu_varref_name"), Some(Value::Str(_)))
-                        && named.contains_key("__mutsu_varref_value") =>
-                {
-                    named.get("__mutsu_varref_value").unwrap_or(arg)
-                }
-                _ => arg,
-            };
-            match arg {
-                Value::Hash(_) | Value::Pair(_, _) => {}
-                Value::Instance {
-                    class_name,
-                    attributes,
-                    ..
-                } if idx == 0 && class_name.resolve() == "IO::Path" => {
-                    first_arg_io_path = true;
-                    if let Some(path) = attributes.as_map().get("path").map(Value::to_string_value)
-                    {
-                        positional.push(path);
-                    } else {
-                        positional.push(arg.to_string_value());
-                    }
-                }
-                Value::Array(elems, _) => {
-                    for elem in elems.iter() {
-                        positional.push(Value::to_string_value(elem));
-                    }
-                }
-                _ => {
-                    positional.push(arg.to_string_value());
-                }
-            }
-        }
-
-        let opts = Self::extract_proc_options(args, 0);
-
-        if positional.is_empty() {
-            return Ok(Self::make_proc_instance(
-                -1,
-                0,
-                0,
-                Value::str(String::new()),
-                opts.capture_err.then(String::new),
-                opts.capture_out.then(String::new),
-            ));
-        }
-
-        let program = &positional[0];
-        let rest_args = &positional[1..];
-
-        // Build the command tuple for .command attribute
-        let command_val = Value::array(positional.iter().map(|s| Value::str(s.clone())).collect());
-        // run() defaults the child's working directory to the dynamic $*CWD
-        // (set by `indir`), not the interpreter process's actual cwd.
-        let opts_cwd = opts
-            .cwd
-            .clone()
-            .or_else(|| self.get_dynamic_string("$*CWD"));
-        let opts_env = opts.env.clone();
-
-        let mut cmd = Command::new(program);
-        Self::apply_run_args(&mut cmd, rest_args, opts.win_verbatim_args);
-
-        // Handle :out with IO::Handle — redirect stdout to that file
-        let mut stdout_file_for_merge: Option<std::fs::File> = None;
-        if let Some(handle_id) = opts.out_handle_id {
-            let table = self.io_handles();
-            let state = table
-                .map
-                .get(&handle_id)
-                .ok_or_else(|| RuntimeError::new("Invalid IO::Handle for :out"))?;
-            let file = state
-                .file
-                .as_ref()
-                .ok_or_else(|| RuntimeError::new(":out IO::Handle has no file"))?;
-            let cloned = file
-                .try_clone()
-                .map_err(|e| RuntimeError::new(format!("Cannot dup file for :out: {e}")))?;
-            if opts.merge {
-                stdout_file_for_merge =
-                    Some(cloned.try_clone().map_err(|e| {
-                        RuntimeError::new(format!("Cannot dup file for :merge: {e}"))
-                    })?);
-            }
-            cmd.stdout(std::process::Stdio::from(cloned));
-        } else if opts.capture_out {
-            cmd.stdout(std::process::Stdio::piped());
-        } else {
-            cmd.stdout(std::process::Stdio::null());
-        }
-        if opts.merge {
-            if let Some(file) = stdout_file_for_merge {
-                cmd.stderr(std::process::Stdio::from(file));
-            } else {
-                // :merge without :out(file) — stderr goes to same pipe as stdout
-                cmd.stderr(std::process::Stdio::piped());
-            }
-        } else if opts.capture_err {
-            cmd.stderr(std::process::Stdio::piped());
-        } else {
-            cmd.stderr(std::process::Stdio::null());
-        }
-        let mut piped_from_live = false;
-        if let Some(src_pid) = opts.in_pipe_pid {
-            let mut child_stdout: Option<std::process::ChildStdout> = None;
-            if let Ok(mut map) = live_proc_map().lock()
-                && let Some(state) = map.get_mut(&src_pid)
-            {
-                child_stdout = state.child.stdout.take();
-            }
-            if let Some(stdout) = child_stdout {
-                cmd.stdin(std::process::Stdio::from(stdout));
-                piped_from_live = true;
-            } else {
-                cmd.stdin(std::process::Stdio::piped());
-            }
-        } else if opts.capture_in || opts.in_pipe_content.is_some() {
-            cmd.stdin(std::process::Stdio::piped());
-        }
-
-        if let Some(cwd) = opts_cwd.clone() {
-            cmd.current_dir(cwd);
-        }
-        for (k, v) in &opts_env {
-            cmd.env(k, v);
-        }
-
-        match cmd.spawn() {
-            Ok(mut child) => {
-                let pid = child.id() as i64;
-
-                if let Some(content) = &opts.in_pipe_content
-                    && let Some(mut stdin) = child.stdin.take()
-                {
-                    use std::io::Write;
-                    let _ = stdin.write_all(content.as_bytes());
-                }
-
-                let needs_live =
-                    (opts.capture_in && opts.in_pipe_content.is_none()) || piped_from_live;
-                if needs_live {
-                    let stdin_handle = child.stdin.take();
-                    // Store child in global map for later access
-                    if let Ok(mut map) = live_proc_map().lock() {
-                        map.insert(
-                            pid,
-                            LiveProcState {
-                                child,
-                                capture_out: opts.capture_out,
-                                capture_err: opts.capture_err,
-                            },
-                        );
-                    }
-                    // Create an IO::Pipe for stdin
-                    let in_pipe = if let Some(stdin) = stdin_handle {
-                        // Store in proc_stdin_map for .print/.close
-                        if let Ok(mut map) = super::native_methods::proc_stdin_map().lock() {
-                            map.insert(pid as u32, Arc::new(std::sync::Mutex::new(Some(stdin))));
-                        }
-                        let mut in_attrs = HashMap::new();
-                        in_attrs.insert("proc-pid".to_string(), Value::Int(pid));
-                        Value::make_instance(Symbol::intern("IO::Pipe"), in_attrs)
-                    } else {
-                        Value::Nil
-                    };
-                    let mut attrs = HashMap::new();
-                    attrs.insert("exitcode".to_string(), Value::Int(-1));
-                    attrs.insert("signal".to_string(), Value::Int(0));
-                    attrs.insert("pid".to_string(), Value::Int(pid));
-                    attrs.insert("command".to_string(), command_val);
-                    attrs.insert("in".to_string(), in_pipe);
-                    attrs.insert("live".to_string(), Value::Bool(true));
-                    if opts.bin {
-                        attrs.insert("bin".to_string(), Value::Bool(true));
-                    }
-                    let proc = Value::make_instance(Symbol::intern("Proc"), attrs);
-                    if let Ok(mut map) = proc_by_pid_map().lock() {
-                        map.insert(pid, proc.clone());
-                    }
-                    return Ok(proc);
-                }
-
-                let captured_out = if opts.capture_out {
-                    child.stdout.take().map(|mut s| {
-                        let mut buf = String::new();
-                        use std::io::Read;
-                        let _ = s.read_to_string(&mut buf);
-                        buf
-                    })
-                } else {
-                    None
-                };
-                let captured_err = if opts.capture_err {
-                    child.stderr.take().map(|mut s| {
-                        let mut buf = String::new();
-                        use std::io::Read;
-                        let _ = s.read_to_string(&mut buf);
-                        buf
-                    })
-                } else {
-                    None
-                };
-                match child.wait() {
-                    Ok(status) => {
-                        let exitcode = status.code().unwrap_or(-1) as i64;
-                        #[cfg(unix)]
-                        let signal = {
-                            use std::os::unix::process::ExitStatusExt;
-                            status.signal().unwrap_or(0) as i64
-                        };
-                        #[cfg(not(unix))]
-                        let signal = 0i64;
-                        Ok(Self::make_proc_instance_bin(
-                            exitcode,
-                            signal,
-                            pid,
-                            command_val,
-                            captured_err,
-                            captured_out,
-                            opts.bin,
-                        ))
-                    }
-                    Err(_) => Ok(Self::make_proc_instance_bin(
-                        -1,
-                        0,
-                        pid,
-                        command_val,
-                        captured_err,
-                        captured_out,
-                        opts.bin,
-                    )),
-                }
-            }
-            Err(err) => {
-                let os_error = err.to_string();
-                // Fallback for cases where $*EXECUTABLE is passed as an IO::Path-ish value
-                // that stringifies ambiguously. Retry with current_exe.
-                if first_arg_io_path || program == "$*EXECUTABLE" || program.ends_with("mutsu") {
-                    let fallback = Some(
-                        Self::resolved_current_executable_path()
-                            .to_string_lossy()
-                            .to_string(),
-                    );
-                    if let Some(exe) = fallback {
-                        let mut retry = Command::new(exe);
-                        Self::apply_run_args(&mut retry, rest_args, opts.win_verbatim_args);
-                        if opts.capture_out {
-                            retry.stdout(std::process::Stdio::piped());
-                        } else {
-                            retry.stdout(std::process::Stdio::null());
-                        }
-                        if opts.capture_err {
-                            retry.stderr(std::process::Stdio::piped());
-                        } else {
-                            retry.stderr(std::process::Stdio::null());
-                        }
-                        if let Some(cwd) = opts_cwd.clone() {
-                            retry.current_dir(cwd);
-                        }
-                        for (k, v) in &opts_env {
-                            retry.env(k, v);
-                        }
-                        if let Ok(mut child) = retry.spawn() {
-                            let pid = child.id() as i64;
-                            let captured_out = if opts.capture_out {
-                                child.stdout.take().map(|mut s| {
-                                    let mut buf = String::new();
-                                    use std::io::Read;
-                                    let _ = s.read_to_string(&mut buf);
-                                    buf
-                                })
-                            } else {
-                                None
-                            };
-                            let captured_err = if opts.capture_err {
-                                child.stderr.take().map(|mut s| {
-                                    let mut buf = String::new();
-                                    use std::io::Read;
-                                    let _ = s.read_to_string(&mut buf);
-                                    buf
-                                })
-                            } else {
-                                None
-                            };
-                            return match child.wait() {
-                                Ok(status) => {
-                                    let exitcode = status.code().unwrap_or(-1) as i64;
-                                    #[cfg(unix)]
-                                    let signal = {
-                                        use std::os::unix::process::ExitStatusExt;
-                                        status.signal().unwrap_or(0) as i64
-                                    };
-                                    #[cfg(not(unix))]
-                                    let signal = 0i64;
-                                    Ok(Self::make_proc_instance(
-                                        exitcode,
-                                        signal,
-                                        pid,
-                                        command_val,
-                                        captured_err,
-                                        captured_out,
-                                    ))
-                                }
-                                Err(_) => Ok(Self::make_proc_instance(
-                                    -1,
-                                    0,
-                                    pid,
-                                    command_val,
-                                    captured_err,
-                                    captured_out,
-                                )),
-                            };
-                        }
-                    }
-                }
-                let mut proc = Self::make_proc_instance(
-                    -1,
-                    0,
-                    0,
-                    command_val,
-                    opts.capture_err.then(String::new),
-                    opts.capture_out.then(String::new),
-                );
-                Self::attach_proc_os_error(&mut proc, &os_error);
-                Ok(proc)
-            }
-        }
-    }
-
-    /// Record the OS-level spawn error on a failed Proc so that
-    /// `X::Proc::Unsuccessful.message` can report it (e.g. command-not-found
-    /// produces "exit code: -1, ... OS error = No such file or directory").
-    fn attach_proc_os_error(proc: &mut Value, os_error: &str) {
-        if let Value::Instance { attributes, .. } = proc {
-            attributes.insert("os-error".to_string(), Value::str(os_error.to_string()));
-        }
-    }
-
-    pub(super) fn builtin_shell(&self, args: &[Value]) -> Result<Value, RuntimeError> {
-        let command_str = args
-            .first()
-            .map(|v| v.to_string_value())
-            .unwrap_or_default();
-        if command_str.is_empty() {
-            return Ok(Self::make_proc_instance(
-                -1,
-                0,
-                0,
-                Value::str(String::new()),
-                None,
-                None,
-            ));
-        }
-
-        let opts = Self::extract_proc_options(args, 1);
-
-        let command_val = Value::str(command_str.clone());
-
-        let mut command = if cfg!(windows) {
-            let mut cmd = Command::new("cmd");
-            cmd.arg("/C").arg(&command_str);
-            cmd
-        } else {
-            let mut cmd = Command::new("sh");
-            cmd.arg("-c").arg(&command_str);
-            cmd
-        };
-
-        if opts.capture_out {
-            command.stdout(std::process::Stdio::piped());
-        } else {
-            command.stdout(std::process::Stdio::null());
-        }
-        if opts.capture_err {
-            command.stderr(std::process::Stdio::piped());
-        } else {
-            command.stderr(std::process::Stdio::null());
-        }
-
-        // shell() defaults the child's working directory to the dynamic $*CWD
-        // (set by `indir`), matching run().
-        if let Some(cwd) = opts
-            .cwd
-            .clone()
-            .or_else(|| self.get_dynamic_string("$*CWD"))
-        {
-            command.current_dir(cwd);
-        }
-        for (k, v) in opts.env {
-            command.env(k, v);
-        }
-
-        match command.spawn() {
-            Ok(mut child) => {
-                let pid = child.id() as i64;
-                let captured_out = if opts.capture_out {
-                    child.stdout.take().map(|mut s| {
-                        let mut buf = String::new();
-                        use std::io::Read;
-                        let _ = s.read_to_string(&mut buf);
-                        buf
-                    })
-                } else {
-                    None
-                };
-                let captured_err = if opts.capture_err {
-                    child.stderr.take().map(|mut s| {
-                        let mut buf = String::new();
-                        use std::io::Read;
-                        let _ = s.read_to_string(&mut buf);
-                        buf
-                    })
-                } else {
-                    None
-                };
-                match child.wait() {
-                    Ok(status) => {
-                        let exitcode = status.code().unwrap_or(-1) as i64;
-                        #[cfg(unix)]
-                        let signal = {
-                            use std::os::unix::process::ExitStatusExt;
-                            status.signal().unwrap_or(0) as i64
-                        };
-                        #[cfg(not(unix))]
-                        let signal = 0i64;
-                        Ok(Self::make_proc_instance(
-                            exitcode,
-                            signal,
-                            pid,
-                            command_val,
-                            captured_err,
-                            captured_out,
-                        ))
-                    }
-                    Err(_) => Ok(Self::make_proc_instance(
-                        -1,
-                        0,
-                        pid,
-                        command_val,
-                        captured_err,
-                        captured_out,
-                    )),
-                }
-            }
-            Err(_) => Ok(Self::make_proc_instance(-1, 0, 0, command_val, None, None)),
-        }
-    }
-
-    pub(super) fn builtin_qx(&mut self, args: &[Value]) -> Result<Value, RuntimeError> {
-        let command = args
-            .first()
-            .map(|v| v.to_string_value())
-            .unwrap_or_default();
-
-        let shell_args = vec![
-            Value::str(command),
-            Value::Pair("out".to_string(), Box::new(Value::Bool(true))),
-            Value::Pair("err".to_string(), Box::new(Value::Bool(true))),
-        ];
-        let proc_value = self.builtin_shell(&shell_args)?;
-
-        if let Value::Instance { attributes, .. } = proc_value {
-            if let Some(err_pipe) = attributes.as_map().get("err") {
-                let stderr = self.call_method_with_values(err_pipe.clone(), "slurp", vec![])?;
-                let stderr_text = stderr.to_string_value();
-                if !stderr_text.is_empty() {
-                    eprint!("{}", stderr_text);
-                    let _ = std::io::Write::flush(&mut std::io::stderr());
-                }
-            }
-            if let Some(out_pipe) = attributes.as_map().get("out") {
-                return self.call_method_with_values(out_pipe.clone(), "slurp", vec![]);
-            }
-        }
-
-        Ok(Value::str(String::new()))
-    }
-
-    pub(super) fn builtin_kill(&self, args: &[Value]) -> Result<Value, RuntimeError> {
-        let signal = args.first().map(super::to_int).unwrap_or(15);
-        let mut success = true;
-        let mut had_pid = false;
-        for val in args.iter().skip(1) {
-            had_pid = true;
-            let pid = super::to_int(val);
-            success &= Self::send_signal(pid, signal);
-        }
-        if !had_pid {
-            success = false;
-        }
-        Ok(Value::Bool(success))
-    }
-
-    pub(super) fn builtin_syscall(&self, args: &[Value]) -> Result<Value, RuntimeError> {
-        let Some(val) = args.first() else {
-            return Ok(Value::Nil);
-        };
-        let num = super::to_int(val);
-        if num == 0 {
-            let pid = self
-                .env
-                .get("*PID")
-                .and_then(|v| match v {
-                    Value::Int(i) => Some(*i),
-                    _ => None,
-                })
-                .unwrap_or_else(|| {
-                    #[cfg(not(target_arch = "wasm32"))]
-                    {
-                        std::process::id() as i64
-                    }
-                    #[cfg(target_arch = "wasm32")]
-                    {
-                        0
-                    }
-                });
-            return Ok(Value::Int(pid));
-        }
-        Ok(Value::Int(-1))
-    }
-
-    pub(super) fn builtin_sleep(&mut self, args: &[Value]) -> Result<Value, RuntimeError> {
-        let secs = Self::seconds_from_value(args.first().cloned());
-        match secs {
-            None | Some(f64::INFINITY) => {
-                // sleep with no args or Inf sleeps indefinitely.
-                // Poll for global exit flag so that `start { exit }` can
-                // wake us up.
-                loop {
-                    thread::sleep(Duration::from_millis(100));
-                    if let Some(code) = super::builtins_control_flow::global_exit_requested() {
-                        self.halted = true;
-                        self.exit_code = code;
-                        return Ok(Value::Nil);
-                    }
-                }
-            }
-            Some(s) => {
-                let duration = Self::duration_from_seconds(Some(s));
-                // For large sleep durations, poll periodically so that
-                // a global exit() from another thread can interrupt us.
-                if duration > Duration::from_secs(10) {
-                    Self::interruptible_sleep(duration);
-                    if let Some(code) = super::builtins_control_flow::global_exit_requested() {
-                        self.halted = true;
-                        self.exit_code = code;
-                        return Ok(Value::Nil);
-                    }
-                } else {
-                    thread::sleep(duration);
-                }
-                self.sync_shared_vars_to_env();
-                Ok(Value::Nil)
-            }
-        }
-    }
-
-    /// Sleep for the given duration, but check for global exit every 100ms.
-    fn interruptible_sleep(total: Duration) {
-        let start = std::time::Instant::now();
-        while start.elapsed() < total {
-            let remaining = total.saturating_sub(start.elapsed());
-            let chunk = remaining.min(Duration::from_millis(100));
-            thread::sleep(chunk);
-            if super::builtins_control_flow::global_exit_requested().is_some() {
-                return;
-            }
-        }
-    }
-
-    pub(super) fn builtin_sleep_timer(&self, args: &[Value]) -> Result<Value, RuntimeError> {
-        let duration = Self::duration_from_seconds(Self::seconds_from_value(args.first().cloned()));
-        let start = Instant::now();
-        thread::sleep(duration);
-        let elapsed = start.elapsed();
-        let remaining = duration.checked_sub(elapsed).unwrap_or_default();
-        Ok(crate::builtins::arith::make_duration_value(
-            remaining.as_secs_f64(),
-        ))
-    }
-
-    pub(super) fn builtin_sleep_till(&self, args: &[Value]) -> Result<Value, RuntimeError> {
-        let arg = args.first().cloned();
-        // Extract the target POSIX timestamp, handling Instant (TAI) values
-        let target_posix = match &arg {
-            Some(Value::Instance {
-                class_name,
-                attributes,
-                ..
-            }) if class_name.resolve() == "Instant" => {
-                // Instant stores TAI time; convert back to POSIX
-                attributes
-                    .as_map()
-                    .get("value")
-                    .and_then(super::to_float_value)
-                    .map(crate::builtins::methods_0arg::temporal::instant_to_posix)
-            }
-            Some(Value::Instance {
-                class_name,
-                attributes,
-                ..
-            }) if class_name.resolve() == "DateTime" => {
-                // Compute posix from DateTime attributes
-                use crate::builtins::methods_0arg::temporal::{datetime_attrs, datetime_to_posix};
-                let (year, month, day, hour, minute, second, timezone) =
-                    datetime_attrs(&(attributes).as_map());
-                Some(datetime_to_posix(
-                    year, month, day, hour, minute, second, timezone,
-                ))
-            }
-            _ => Self::seconds_from_value(arg),
-        };
-        if let Some(target) = target_posix {
-            let now = crate::value::current_time_secs_f64();
-            if target <= now {
-                return Ok(Value::Bool(false));
-            }
-            let diff_secs = target - now;
-            thread::sleep(Duration::from_secs_f64(diff_secs));
-            return Ok(Value::Bool(true));
-        }
-        Ok(Value::Bool(false))
-    }
-
-    /// `start { ... }` — spawn a thread to execute the block and return a Promise.
-    pub(super) fn builtin_start(&mut self, args: Vec<Value>) -> Result<Value, RuntimeError> {
-        let block = args.into_iter().next().unwrap_or(Value::Nil);
-        Ok(self.spawn_callable_promise(block, Symbol::intern("Promise")))
-    }
-
-    /// `await` — block until Promise(s) resolve, then return their results.
-    /// Build the `RuntimeError` carrying an `X::Await::Died` exception for a
-    /// broken Promise observed by `await`. Raku mixes the `X::Await::Died` role
-    /// INTO the original failure cause (`.^name` is e.g. `X::AdHoc+{X::Await::Died}`),
-    /// so it keeps its class, message and backtrace while also doing
-    /// `X::Await::Died`. We mirror that: when the cause is an exception instance,
-    /// mark it with `__mutsu_does_await_died` (recognized by `isa_check`) and
-    /// re-raise it unchanged, preserving its gist/backtrace. A plain (non-instance)
-    /// cause is wrapped in a fresh `X::Await::Died`.
-    /// Format the current routine stack (the point where `await` was called) as
-    /// a Raku-style backtrace string, innermost frame first. Used to append the
-    /// await/re-throw location to a broken-Promise exception's backtrace.
-    fn await_site_backtrace(&self) -> String {
-        let stack = self.routine_stack();
-        let mut lines = Vec::new();
-        for frame in stack.iter().rev() {
-            let loc = match (frame.file.as_deref(), frame.line) {
-                (Some(f), Some(l)) => format!(" at {} line {}", f, l),
-                (Some(f), None) => format!(" at {}", f),
-                _ => String::new(),
-            };
-            if frame.is_block || frame.name.is_empty() || frame.name == "<unit>" {
-                lines.push(format!("  in block <unit>{}", loc));
-            } else {
-                lines.push(format!("  in sub {}{}", frame.name, loc));
-            }
-        }
-        lines.join("\n")
-    }
-
-    fn await_died_error(&self, cause: Value) -> RuntimeError {
-        // Append the await call-site to the exception's backtrace, so the gist
-        // surfaces both where the Promise's code died (preserved) and where it
-        // was awaited (the re-throw location) — matching Raku's X::Await::Died.
-        if let Value::Instance { attributes, .. } = &cause {
-            let await_bt = self.await_site_backtrace();
-            if !await_bt.is_empty()
-                && let Some(Value::Instance {
-                    class_name: bt_cn,
-                    attributes: bt_attrs,
-                    ..
-                }) = attributes.as_map().get("backtrace").cloned()
-                && bt_cn == "Backtrace"
-            {
-                let old = bt_attrs
-                    .as_map()
-                    .get("text")
-                    .map(Value::to_string_value)
-                    .unwrap_or_default();
-                let combined = if old.is_empty() {
-                    await_bt
-                } else {
-                    format!("{old}\n{await_bt}")
-                };
-                bt_attrs.insert("text".to_string(), Value::str(combined));
-            }
-        }
-        let msg = match &cause {
-            Value::Instance { attributes, .. } => {
-                attributes.insert("__mutsu_does_await_died".to_string(), Value::Bool(true));
-                attributes
-                    .as_map()
-                    .get("message")
-                    .map(Value::to_string_value)
-                    .unwrap_or_else(|| cause.to_string_value())
-            }
-            other => other.to_string_value(),
-        };
-        let mut err = RuntimeError::new(msg.clone());
-        let exc = if matches!(cause, Value::Instance { .. }) {
-            cause
-        } else {
-            let mut attrs = std::collections::HashMap::new();
-            attrs.insert("message".to_string(), Value::str(msg.clone()));
-            attrs.insert("payload".to_string(), Value::str(msg));
-            attrs.insert("__mutsu_does_await_died".to_string(), Value::Bool(true));
-            Value::make_instance(Symbol::intern("X::Await::Died"), attrs)
-        };
-        err.exception = Some(Box::new(exc));
-        err
-    }
-
-    /// Normalize a single `await` target. A `Supply` is awaited via its
-    /// `.Promise` (kept with the last emitted value when the supply completes,
-    /// broken if it quits) — `await $supply` ≡ `await $supply.Promise`. All other
-    /// values pass through unchanged (Promises and Channels are handled inline).
-    fn await_normalize(&mut self, v: Value) -> Result<Value, RuntimeError> {
-        if let Value::Instance { class_name, .. } = &v
-            && class_name == "Supply"
-        {
-            return self.call_method_with_values(v, "Promise", vec![]);
-        }
-        Ok(v)
-    }
-
-    pub(super) fn builtin_await(&mut self, args: &[Value]) -> Result<Value, RuntimeError> {
-        if args.is_empty() {
-            return Err(RuntimeError::new(
-                "Must specify a Promise or list of Promises to await",
-            ));
-        }
-        let mut await_targets: Vec<Value> = Vec::new();
-        for arg in args {
-            if let Some(items) = arg.as_list_items() {
-                let items: Vec<Value> = items.to_vec();
-                for it in items {
-                    await_targets.push(self.await_normalize(it)?);
-                }
-            } else {
-                let normalized = self.await_normalize(arg.clone())?;
-                await_targets.push(normalized);
-            }
-        }
-        let mut results = Vec::new();
-        for arg in &await_targets {
-            match arg {
-                Value::Promise(shared) => {
-                    let (result, output, stderr) = shared.wait();
-                    self.emit_output(&output);
-                    self.output_sink_mut().stderr_output.push_str(&stderr);
-                    if let Some(payload) = shared.take_thread_payload()
-                        && let Ok(payload) = payload.downcast::<ThreadPromisePayload>()
-                    {
-                        let ThreadPromisePayload {
-                            new_handles,
-                            next_handle_id,
-                        } = *payload;
-                        let mut table = self.io_handles_mut();
-                        for (id, state) in new_handles {
-                            table.map.entry(id).or_insert(state);
-                        }
-                        if next_handle_id > table.next_id {
-                            table.next_id = next_handle_id;
-                        }
-                    }
-                    if shared.status() == "Broken" {
-                        self.sync_shared_vars_to_env();
-                        // Raku wraps the broken Promise's cause in X::Await::Died.
-                        return Err(self.await_died_error(result));
-                    }
-                    // Replay deferred Proc::Async taps
-                    if let Value::Instance {
-                        ref class_name,
-                        ref attributes,
-                        ..
-                    } = result
-                        && class_name == "Proc"
-                    {
-                        self.replay_proc_taps(attributes);
-                    }
-                    let result = Self::unwrap_async_status_result(result)?;
-                    results.push(result);
-                    // Sync shared variables after each thread completes
-                    self.sync_shared_vars_to_env();
-                }
-                // Backward compat: Instance-based Promise
-                Value::Instance {
-                    class_name,
-                    attributes,
-                    ..
-                } if class_name == "Promise" => {
-                    let result = attributes
-                        .as_map()
-                        .get("result")
-                        .cloned()
-                        .unwrap_or(Value::Nil);
-                    results.push(result);
-                }
-                Value::Array(elems, ..) => {
-                    for elem in elems.iter() {
-                        match elem {
-                            Value::Promise(shared) => {
-                                let (result, output, stderr) = shared.wait();
-                                self.emit_output(&output);
-                                self.output_sink_mut().stderr_output.push_str(&stderr);
-                                if shared.status() == "Broken" {
-                                    self.sync_shared_vars_to_env();
-                                    return Err(self.await_died_error(result));
-                                }
-                                let result = Self::unwrap_async_status_result(result)?;
-                                results.push(result);
-                            }
-                            Value::Instance {
-                                class_name,
-                                attributes,
-                                ..
-                            } if class_name == "Promise" => {
-                                let result = attributes
-                                    .as_map()
-                                    .get("result")
-                                    .cloned()
-                                    .unwrap_or(Value::Nil);
-                                results.push(result);
-                            }
-                            _ => results.push(elem.clone()),
-                        }
-                    }
-                }
-                // `await $channel` blocks for the next value (like `.receive`):
-                // returns it, or throws X::Channel::ReceiveOnClosed when the
-                // channel is drained and closed, or the failure cause when failed.
-                Value::Channel(ch) => match ch.receive_result() {
-                    Ok(Value::Nil) => return Err(Self::channel_receive_closed_error()),
-                    Ok(value) => results.push(value),
-                    Err(cause) => {
-                        let ex = Self::as_exception_value(cause);
-                        let mut err = RuntimeError::new(ex.to_string_value());
-                        err.exception = Some(Box::new(ex));
-                        return Err(err);
-                    }
-                },
-                _ => results.push(arg.clone()),
-            }
-        }
-        // Sync shared variables back from child threads
-        self.sync_shared_vars_to_env();
-
-        // Process any pending instance destroys (e.g. objects that were nilled
-        // before the start block but whose DESTROY wasn't yet triggered).
-        self.closure_env_overrides.clear();
-        self.run_pending_instance_destroys()?;
-
-        // Drain shared thread output buffers (concurrent output interleaved in real order).
-        // Clone the Arc out so the output_sink guard is dropped before emit_output re-borrows self.
-        let shared_out = self.output_sink().shared_thread_output.clone();
-        if let Some(shared) = shared_out {
-            let drained = std::mem::take(&mut *shared.lock().unwrap());
-            if !drained.is_empty() {
-                self.emit_output(&drained);
-            }
-        }
-        let shared_err = self.output_sink().shared_thread_stderr.clone();
-        if let Some(shared) = shared_err {
-            let drained = std::mem::take(&mut *shared.lock().unwrap());
-            if !drained.is_empty() {
-                self.output_sink_mut().stderr_output.push_str(&drained);
-            }
-        }
-
-        // Slip results flatten into the surrounding await result list (slip
-        // semantics): `await start { (2, 3).Slip }, start { 4 }` yields
-        // `(2, 3, 4)`, not `((2, 3), 4)`. A plain List/Array result is NOT a
-        // Slip and stays nested, matching Raku.
-        if results.iter().any(|r| matches!(r, Value::Slip(_))) {
-            let mut flat = Vec::with_capacity(results.len());
-            for r in results {
-                if let Value::Slip(items) = r {
-                    flat.extend(items.iter().cloned());
-                } else {
-                    flat.push(r);
-                }
-            }
-            results = flat;
-        }
-
-        if results.len() == 1 {
-            Ok(results.into_iter().next().unwrap())
-        } else {
-            Ok(Value::array(results))
-        }
-    }
-
-    fn unwrap_async_status_result(value: Value) -> Result<Value, RuntimeError> {
-        if let Value::Instance {
-            class_name,
-            attributes,
-            ..
-        } = &value
-            && class_name == "IO::Socket::Async::StatusResult"
-        {
-            let status = attributes
-                .as_map()
-                .get("status")
-                .map(Value::to_string_value)
-                .unwrap_or_else(|| "Broken".to_string());
-            if status == "Kept" {
-                return Ok(attributes
-                    .as_map()
-                    .get("result")
-                    .cloned()
-                    .unwrap_or(Value::Nil));
-            }
-            let cause = attributes
-                .as_map()
-                .get("cause")
-                .cloned()
-                .unwrap_or(Value::Nil);
-            let message = cause.to_string_value();
-            let mut err = RuntimeError::new(message.clone());
-            if matches!(cause, Value::Instance { .. }) {
-                err.exception = Some(Box::new(cause));
-            } else {
-                let mut attrs = std::collections::HashMap::new();
-                attrs.insert("payload".to_string(), Value::str(message.clone()));
-                attrs.insert("message".to_string(), Value::str(message));
-                let ex = Value::make_instance(Symbol::intern("X::AdHoc"), attrs);
-                err.exception = Some(Box::new(ex));
-            }
-            return Err(err);
-        }
-        Ok(value)
-    }
-    /// `signal(SIGINT, ...)` — returns a Supply that emits Signal enum values
-    /// when the process receives the specified OS signals.
-    pub(super) fn builtin_signal(&mut self, args: &[Value]) -> Result<Value, RuntimeError> {
-        use std::sync::mpsc;
-
-        // Extract signal numbers and their enum representations
-        let signals: Vec<(i64, Value)> = args
-            .iter()
-            .filter_map(|v| match v {
-                Value::Enum { value, .. } => Some((value.as_i64(), v.clone())),
-                Value::Int(i) => Some((*i, v.clone())),
-                _ => None,
-            })
-            .collect();
-
-        let supply_id = super::native_methods::next_supply_id();
-
-        // Create channel for the Supply
-        let (tx, rx) = mpsc::channel();
-
-        // Register the channel in the supply channel map
-        if let Ok(mut map) = super::native_methods::supply_channel_map_pub().lock() {
-            map.insert(supply_id, rx);
-        }
-
-        // Set up real signal handling using pipe + sigaction
-        for (signum, sig_val) in &signals {
-            signal_watcher::register_signal(*signum as i32, supply_id, tx.clone(), sig_val.clone());
-        }
-
-        let mut attrs = std::collections::HashMap::new();
-        attrs.insert("values".to_string(), Value::array(Vec::new()));
-        attrs.insert("taps".to_string(), Value::array(Vec::new()));
-        attrs.insert("supply_id".to_string(), Value::Int(supply_id as i64));
-        Ok(Value::make_instance(Symbol::intern("Supply"), attrs))
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn extract_proc_options_reads_win_verbatim_pair() {
-        let args = vec![Value::Pair(
-            "win-verbatim-args".to_string(),
-            Box::new(Value::Bool(true)),
-        )];
-        let opts = Interpreter::extract_proc_options(&args, 0);
-        assert!(opts.win_verbatim_args);
-    }
-
-    #[test]
-    fn extract_proc_options_reads_win_verbatim_hash() {
-        let mut map = HashMap::new();
-        map.insert("win-verbatim-args".to_string(), Value::Bool(true));
-        let args = vec![Value::hash(map)];
-        let opts = Interpreter::extract_proc_options(&args, 0);
-        assert!(opts.win_verbatim_args);
     }
 }
