+++ src/runtime/mod.rs
@@ -152,6 +152,10 @@ mod builtins_operators;
 pub(crate) mod builtins_reduce;
 mod builtins_string;
 mod builtins_system;
+mod builtins_system_async;
+mod builtins_system_proc;
+mod builtins_system_require;
+mod builtins_system_run;
 mod builtins_unbase;
 mod call_helpers;
 mod calls;
